Apply improved types & architecture to Tron
